add coupon & edit coupon functionality

+++ src/app/admin/_actions/coupons.tsx
@@ -5,27 +5,27 @@ import { z } from "zod"
 import fs from "fs/promises"
 import { revalidatePath } from "next/cache"
 
-const fileSchema = z.instanceof(File, { message: "required" })
-    .refine(file => file.size === 0 || file.type.startsWith('image/'), { message: "File must be an image" });
+const fileSchema = z.instanceof(File)
+  .refine(file => file.size === 0 || file.type.startsWith('image/'), { message: "File must be an image" });
 
 const addSchema = z.object({
-    name: z.string().min(1),
-    code: z.string().min(5),
+    code: z.string().min(3),
     description: z.string().optional(),
     maxDiscount: z.coerce.number().int().min(1),
     minOrderValue: z.coerce.number().int().min(1),
-    image: fileSchema.optional().refine(file => file == null || file.size > 0, 'required'),
+    image: fileSchema.optional()
 });
 
 export async function addCoupon(preState: unknown, formData: FormData) {
     const result = addSchema.safeParse(Object.fromEntries(formData.entries()))
     console.log(Object.fromEntries(formData.entries()))
     if (result.success === false) {
+        console.log("getting error in coupon",result.error)
         return result.error.formErrors.fieldErrors
     }
 
     const coupon = result.data
-
+    console.log("coupon data: ",coupon)
     let filePath;
     if(coupon.image){
         await fs.mkdir("public/coupons", { recursive: true })
@@ -65,7 +65,9 @@ export async function updateCoupon(id: string, preState: unknown, formData: Form
 
     let filePath = coupon.image
     if (data.image != null && data.image.size > 0) {
-        await fs.unlink(`public${coupon.image}`)
+        if(coupon.image){
+            await fs.unlink(`public${coupon.image}`)
+        }
         filePath = `/coupons/${crypto.randomUUID()}-${data.image.name}`
         await fs.writeFile(`public${filePath}`, Buffer.from(await data.image.arrayBuffer()))
     }
